// packages/notation-model/src/mnx/types.ts
export type MnxDocument = GlobalAttrs & {
  global: Global;
  layouts?: SystemLayout[];
  mnx: Mnx;
  parts: Part[];
  scores?: Score[];
};
export type String = string;
export type Id = string;
export type Global = GlobalAttrs & {
  lyrics?: LyricsGlobal;
  measures: MeasureGlobal[];
  sounds?: SoundsGlobal;
};
export type LyricsGlobal = GlobalAttrs & {
  lineMetadata?: LyricLinesMetadata;
  lineOrder?: LyricLineId[];
};
/**
 * This interface was referenced by `LyricLinesMetadata`'s JSON-Schema definition
 * via the `patternProperty` "^.*$".
 */
export type LyricLineMetadata = GlobalAttrs & {
  label?: LyricLineLabel;
  lang?: LanguageCode;
};
export type LyricLineLabel = string;
export type LanguageCode = string;
export type LyricLineId = string;
export type MeasureGlobal = GlobalAttrs & {
  barline?: Barline;
  ending?: Ending;
  fermata?: Fermata;
  fine?: Fine;
  jump?: Jump;
  key?: Key;
  number?: MeasureNumber;
  repeatEnd?: RepeatEnd;
  repeatStart?: RepeatStart;
  segno?: Segno;
  tempos?: Tempo[];
  time?: Time;
};
// ...
export type SystemLayout = GlobalAttrs & {
  content: SystemLayoutContent;
};
// ...
export type SystemLayoutContent = (StaffGroup | Staff)[];
export type Mnx = GlobalAttrs & {
  support?: Support;
  version: VersionNumber;
};
export type Support = GlobalAttrs & {
  useAccidentalDisplay?: boolean;
  useBeams?: boolean;
};
export type VersionNumber = number;
export type Part = GlobalAttrs & {
  kit?: Kit;
  measures: PartMeasure[];
  name?: PartName;
  shortName?: PartShortName;
  smuflFont?: SmuflFont;
  staves?: StaffCount;
  transposition?: PartTransposition;
};
/**
 * This interface was referenced by `Kit`'s JSON-Schema definition
 * via the `patternProperty` "^.*$".
 */
export type KitComponent = GlobalAttrs & {
  name?: String;
  sound?: Id;
  staffPosition: StaffPosition;
};
export type StaffPosition = number;
export type PartMeasure = GlobalAttrs & {
  arpeggios?: Arpeggio[];
  beams?: BeamList;
  clefs?: PositionedClef[];
  dynamics?: DynamicGroup[];
  measureRepeat?: MeasureRepeat;
  nonArpeggios?: NonArpeggio[];
  ottavas?: Ottava[];
  sequences: Sequence[];
  staffConfigs?: PositionedStaffConfig[];
};
// ...
export type Pitch = GlobalAttrs & {
  alter?: Alter;
  octave: Octave;
  step: Step;
};
export type Alter = number;
export type Octave = number;
export type Step = 'A' | 'B' | 'C' | 'D' | 'E' | 'F' | 'G';
// ...
export type PartName = string;
export type PartShortName = string;
export type SmuflFont = string;
export type StaffCount = number;
export type PartTransposition = GlobalAttrs & {
  interval: Interval;
  keyFifthsFlipAt?: IntegerSigned;
  prefersWrittenPitches?: boolean;
};
export type Interval = GlobalAttrs & {
  halfSteps: IntegerSigned;
  staffDistance: IntegerSigned;
};
export type Score = GlobalAttrs & {
  layout?: Id;
  multimeasureRests?: MultimeasureRest[];
  name: ScoreName;
  pages?: Page[];
  useWritten?: boolean;
};
export type MultimeasureRest = GlobalAttrs & {
  duration: MeasureCount;
  label?: String;
  start: Id;
};
export type MeasureCount = number;
export type ScoreName = string;
export type Page = GlobalAttrs & {
  layout?: Id;
  systems: System[];
};
export type System = GlobalAttrs & {
  layout?: Id;
  layoutChanges?: LayoutChange[];
  measure: Id;
};
export type LayoutChange = GlobalAttrs & {
  layout: Id;
  location: MeasureRhythmicPosition;
};

export interface GlobalAttrs {
  _c?: String;
  _x?: VendorExtensions;
  id?: Id;
}
export interface VendorExtensions {
  [k: string]: VendorDict;
}
/**
 * This interface was referenced by `VendorExtensions`'s JSON-Schema definition
 * via the `patternProperty` "^.*$".
 */
export interface VendorDict {}
export interface LyricLinesMetadata {
  [k: string]: LyricLineMetadata;
}
export interface SoundsGlobal {
  [k: string]: Sound;
}
export interface Kit {
  [k: string]: KitComponent;
}